+++ apps/app-backend/src/functions/__tests__/featured-institutions.test.ts
@@ -1,6 +1,7 @@
-import * as azureStorage from "@pagopa/io-functions-commons/dist/src/utils/azure_storage";
-import { BlobService } from "azure-storage";
+import * as blobStorageClientHelper from "../../utils/blob-storage/helper";
+import { BlobServiceClient } from "@azure/storage-blob";
 import * as E from "fp-ts/lib/Either";
+import * as TE from "fp-ts/lib/TaskEither";
 import * as O from "fp-ts/lib/Option";
 import { describe, expect, it, vi } from "vitest";
 import { IConfig } from "../../config";
@@ -10,10 +11,10 @@ import { makeFeaturedInstitutionsHandler } from "../featured-institutions";
 import { Institutions } from "../../generated/definitions/internal/Institutions";
 
 // blobService Base Mock
-const mockBlobService = {} as unknown as BlobService;
+const mockBlobServiceClient = {} as unknown as BlobServiceClient;
 const mockUpsertBlobFromObject = vi
-  .spyOn(azureStorage, "getBlobAsObject")
-  .mockResolvedValue(E.right(O.some(mockFeaturedInstitutions)));
+  .spyOn(blobStorageClientHelper, "getBlobAsObject")
+  .mockReturnValue(TE.right(O.some(mockFeaturedInstitutions)));
 
 const mockedConfiguration = {
   FEATURED_ITEMS_CONTAINER_NAME: "container",
@@ -24,13 +25,13 @@ describe("Get Featured Institutions", () => {
   it("should return featured institutions", async () => {
     const result = await makeFeaturedInstitutionsHandler(mockedConfiguration)({
       ...httpHandlerInputMocks,
-      blobService: mockBlobService,
+      blobServiceClient: mockBlobServiceClient,
     })();
 
     expect(mockUpsertBlobFromObject).toBeCalled();
     expect(mockUpsertBlobFromObject).toBeCalledWith(
       Institutions,
-      mockBlobService,
+      mockBlobServiceClient,
       mockedConfiguration.FEATURED_ITEMS_CONTAINER_NAME,
       mockedConfiguration.FEATURED_INSTITUTIONS_FILE_NAME
     );
@@ -46,19 +47,19 @@ describe("Get Featured Institutions", () => {
 
   it("should return internal error", async () => {
     const errorMessage = "Error blobService";
-    mockUpsertBlobFromObject.mockResolvedValueOnce(
-      E.left(new Error(errorMessage))
+    mockUpsertBlobFromObject.mockReturnValueOnce(
+      TE.left(new Error(errorMessage))
     );
 
     const result = await makeFeaturedInstitutionsHandler(mockedConfiguration)({
       ...httpHandlerInputMocks,
-      blobService: mockBlobService,
+      blobServiceClient: mockBlobServiceClient,
     })();
 
     expect(mockUpsertBlobFromObject).toBeCalled();
     expect(mockUpsertBlobFromObject).toBeCalledWith(
       Institutions,
-      mockBlobService,
+      mockBlobServiceClient,
       mockedConfiguration.FEATURED_ITEMS_CONTAINER_NAME,
       mockedConfiguration.FEATURED_INSTITUTIONS_FILE_NAME
     );
@@ -76,17 +77,17 @@ describe("Get Featured Institutions", () => {
   });
 
   it("should return an empty list when file is not found", async () => {
-    mockUpsertBlobFromObject.mockResolvedValueOnce(E.right(O.none));
+    mockUpsertBlobFromObject.mockReturnValueOnce(TE.right(O.none));
 
     const result = await makeFeaturedInstitutionsHandler(mockedConfiguration)({
       ...httpHandlerInputMocks,
-      blobService: mockBlobService,
+      blobServiceClient: mockBlobServiceClient,
     })();
 
     expect(mockUpsertBlobFromObject).toBeCalled();
     expect(mockUpsertBlobFromObject).toBeCalledWith(
       Institutions,
-      mockBlobService,
+      mockBlobServiceClient,
       mockedConfiguration.FEATURED_ITEMS_CONTAINER_NAME,
       mockedConfiguration.FEATURED_INSTITUTIONS_FILE_NAME
     );
