+++ apps/app-backend/src/utils/blob-storage/dependency.ts
@@ -1,9 +1,4 @@
-import { BlobService } from "azure-storage";
 import { BlobServiceClient } from "@azure/storage-blob";
-
-export type BlobServiceDependency = {
-  readonly blobService: BlobService;
-};
 
 export type BlobServiceClientDependency = {
   readonly blobServiceClient: BlobServiceClient;
